[FEAT]: Bookmarks structure.
