+++ src/views/Bookmarks/Bookmarks.jsx
@@ -1,7 +1,8 @@
 import React, { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import BookmarksList from "@components/BookmarksList/BookmarksList";
 import { getFavorites } from "@redux/actions/Favorites/getFavorites";
+import BookmarkList from "@components/BookmarksList/BookmarkList";
+import FeaturedContainer from "@components/FeaturedContainer/FeaturedContainer";
 
 export default function Bookmarks() {
 
@@ -19,11 +20,17 @@ export default function Bookmarks() {
 
     // COMPONENT:
     return (
-        <main className="flex flex-col w-full h-full p-whiteSpaceTop bg-bg">
-            <section className="flex items-center justify-center w-full mt-[50px] bg-bg">
+        <main className="flex flex-col w-full h-full p-whiteSpaceTop mb-[50px] bg-bg">
+            <section className="flex items-center justify-center w-full bg-bg">
                 <div className="flex flex-col justify-center m-sides max-w-maxSc w-maxIn">
                     <h1 className="text-[50px] font-bold text-left uppercase">Productos guardados</h1>
-                    <BookmarksList favorites={favorites} />
+                    <BookmarkList favorites={favorites} />
+                </div>
+            </section>
+            <section className="flex items-center justify-center w-full mt-[50px] bg-bg">
+                <div className="flex flex-col justify-center m-sides max-w-maxSc w-maxIn">
+                    <h2>Productos recomendados</h2>
+                    <FeaturedContainer bestSellersContainer={true} />
                 </div>
             </section>
         </main>
